5.0-python/2.1-flask-todo/requirements.txt
Flask==1.1.2
flask-bootstrap-3.3.7.1
flask-wtf-0.14.3
flask-testing-0.8.1
blinker-1.4
Pyrebase==3.0.27
